feat: after auth, add & view self posted data

+++ api/auth.py
@@ -6,7 +6,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.authentication import TokenAuthentication, SessionAuthentication
 from rest_framework import status
 # from base.models import User
-from .serializers import UserSerializer
+from .serializers import UserSerializer_SignUp, UserSerializer_SignIn, UserSerializer_Authed
 from django.core import serializers
 
 def authenticate_user(email, password):
@@ -22,12 +22,9 @@ def authenticate_user(email, password):
     return None
 
 @api_view(['GET'])
-@authentication_classes([SessionAuthentication, TokenAuthentication])
-@permission_classes([IsAuthenticated])
 def get_data(request):
-    permission_classes = [IsAuthenticated]
-    users = User.objects.all()
-    serializer = UserSerializer(users, many = True)
+    authed_user = TokenAuthentication().authenticate(request)
+    serializer = UserSerializer_Authed(authed_user[0])
     return Response(serializer.data)
 
 
@@ -35,12 +32,17 @@ def get_data(request):
 @authentication_classes([])
 @permission_classes([])
 def add_user(request):
-    serializer = UserSerializer(data = request.data)
+    serializer = UserSerializer_SignUp(data = request.data)
     is_valid = serializer.is_valid()
     errors = serializer.errors
-    name = request.data.get('name')
+    username = request.data.get('username')
     email = request.data.get('email')
     password = request.data.get('password')
+    first_name = ''
+    last_name = ''
+    if len(username.split(' ')) > 1:
+        first_name = username.split(' ')[0]
+        last_name = username.split(' ')[1]
     user = authenticate_user(email, password)
     if user is not None:
         response = {
@@ -50,7 +52,10 @@ def add_user(request):
     else:
         try:
             if is_valid:
-                user= User.objects.create_user(name, email, password)
+                user = User.objects.create_user(username, email, password)
+                user.first_name = first_name
+                user.last_name = last_name
+                user.save()
                 token = Token.objects.create(user=user)
                 response = {
                     "message": "Added successfully",
@@ -74,26 +79,40 @@ def add_user(request):
 @authentication_classes([])
 @permission_classes([])
 def auth_user(request):
-    serializer = UserSerializer(data = request.data)
-    serializer.is_valid()
+    serializer = UserSerializer_SignIn(data = request.data)
+    is_valid = serializer.is_valid()
+    errors = serializer.errors
     email = serializer.data.get('email')
     password = serializer.data.get('password')
     user = authenticate_user(email, password)
-    # user = User.objects.get(email=email)
-    print('user: ' + str(user))
-    if user is not None:
-        Token.objects.filter(user=user).delete()
-        token = Token.objects.create(user=user)
+    try:
+        if is_valid:
+            if user is not None:
+                Token.objects.filter(user=user).delete()
+                token = Token.objects.create(user=user)
+                response = {
+                    "message": "Authenticated successfully",
+                    "token": token.key
+                }
+                return Response(response, status=status.HTTP_200_OK)
+            else:
+                response = {
+                    "message": "Incorrect email or password"
+                }
+                return Response(response, status=status.HTTP_409_CONFLICT)
+        else :
+            response = {
+                "message": "An error occurred",
+                "error": errors
+            }
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+            
+    except Exception as e:
         response = {
-            "message": "Authenticated successfully",
-            "token": token.key
+            "message": "An error occured",
+            "error": str(e)
         }
-        return Response(response, status=status.HTTP_200_OK)
-    else:
-        response = {
-            "message": "Incorrect email or password"
-        }
-        return Response(response, status=status.HTTP_409_CONFLICT)
+        return Response(response, status=status.HTTP_400_BAD_REQUEST)
